Renamed DhalangPDF to pdf and added it to dhalang namespace
