+++ spec/PDF_spec.rb
@@ -1,31 +1,31 @@
 require 'rspec'
-require 'DhalangPDF'
+require 'PDF'
 require 'pdf/reader'
 
 # Method: DhalangPDF.get_from_url
 describe '#get_from_url' do
   context 'url without specified protocol' do
     it 'should raise InvalidURIError' do
-      expect { DhalangPDF.get_from_url("google.com") }.to raise_error(URI::InvalidURIError)
+      expect { Dhalang::PDF.get_from_url("google.com") }.to raise_error(URI::InvalidURIError)
     end
   end
 
   context 'valid url' do
     it 'should not raise ArgumentError' do
-      expect { DhalangPDF.get_from_url("https://www.google.com") }.to_not raise_error(URI::InvalidURIError)
+      expect { Dhalang::PDF.get_from_url("https://www.google.com") }.to_not raise_error(URI::InvalidURIError)
     end
 
     it 'should return an object of type string' do
-      expect(DhalangPDF.get_from_url("https://www.google.com")).to be_an_instance_of(String)
+      expect(Dhalang::PDF.get_from_url("https://www.google.com")).to be_an_instance_of(String)
     end
 
     it 'should return a file that is not empty' do
-      pdf_binary_content = DhalangPDF.get_from_url("https://www.google.com")
+      pdf_binary_content = Dhalang::PDF.get_from_url("https://www.google.com")
       expect(File.zero?(create_pdf_file(pdf_binary_content))).to be false
     end
 
     it 'should return pdf containing the webpage' do
-      pdf_binary_content = DhalangPDF.get_from_url("https://www.google.com")
+      pdf_binary_content = Dhalang::PDF.get_from_url("https://www.google.com")
       pdf_reader = PDF::Reader.new(create_pdf_file(pdf_binary_content).path)
       expect(pdf_reader.page(1).to_s.include?("Google")).to be true
     end
@@ -37,7 +37,7 @@ describe '#get_from_html' do
   context 'invalid html' do
     it 'should return empty pdf' do
       html = ""
-      pdf_binary_content = DhalangPDF.get_from_html(html)
+      pdf_binary_content = Dhalang::PDF.get_from_html(html)
       pdf_reader = PDF::Reader.new(create_pdf_file(pdf_binary_content).path)
       expect(pdf_reader.page(1).to_s.empty?).to be true
     end
@@ -46,7 +46,7 @@ describe '#get_from_html' do
   context 'valid html' do
     it 'should return pdf containing the html' do
       html = "<html><head></head><body><h1>hello world</h1></body></html>"
-      pdf_binary_content = DhalangPDF.get_from_html(html)
+      pdf_binary_content = Dhalang::PDF.get_from_html(html)
       pdf_reader = PDF::Reader.new(create_pdf_file(pdf_binary_content).path)
       expect(pdf_reader.page(1).to_s.include?("hello world")).to be true
     end
